# Optimized physics object spawn time