- Optimized physics object spawn time
- Optimized shape usage when the shape is not scaled

fixes: #30027

The old bullet integration was 2 times slower compared to Godot physics when adding new bodies.
With this PR the timings got improved drastically (in the order of 1000 times).

(_All measurements were done with release debug_)

### Before optimization:
Bullet Physics Add 4000 Body time: 66 seconds
Godot Physics Add 4000 Body time: 28 seconds

### After optimization:
Bullet Physics Add 4000 Body time: ~0.07 seconds
Godot Physics Add 4000 Body time: 28 seconds

Bullet Physics Add 100000 Body time: ~6.5 seconds
Godot Physics Add 100000 Body time: **Crash after ~4500 bodies**

The tests got executed using:
```
extends Node

var shapes_per_frame := 4000
var shape := BoxShape3D.new()
var t1

func _ready():
	t1 = OS.get_unix_time()
	_do_test()


func _physics_process(delta):
	call_deferred("quit")


func quit():
	# Quit next frame.
	var t2 = OS.get_unix_time()
	print("Done in: ", t2 - t1)
	get_tree().quit()


func _do_test():
	print("Test start")
	for i in range(0, shapes_per_frame):
		var rb = RigidBody3D.new()
		var c = CollisionShape3D.new()
		c.set_shape(shape)
		rb.add_child(c)
		add_child(rb)
		if i % 50 == 0:
			print("Added: ", i)

```