@asheraryam 2020-06-21
I took a look at the changes and none of it seems to depend on 4.0 changes, so I believe this can apply to 3.2.x as well.

After this is merged I can port it to 3.2 if it would be welcome.