@aaronfranke 2020-06-22
This is a massive improvement, thank you for making this.

For testing, I [rebased this on the 3.2 branch](https://github.com/aaronfranke/godot/tree/3.2_add_body_impr_physics) and opened the [Voxel demo project](https://github.com/godotengine/godot-demo-projects/tree/master/3d/voxel) (but with VSync turned off). Each frame the demo generates 1 chunk. As the world generates, the FPS gradually becomes lower and lower.

* 3.2 branch: At render distance 6, min FPS is around 34. At render distance 10, min FPS is around 7.

* With [this commit on 3.2](https://github.com/aaronfranke/godot/tree/3.2_add_body_impr_physics): At render distance 6, min FPS is around 127. At render distance 10, min FPS is around 14.

So the performance in my case is significantly increased, and the gains are exponentially higher when there are fewer chunks in total (and therefore the slowness mostly comes from generating new ones rather than processing existing ones).

Also, the above is comparing a debug build with a debug build. With 3.2 release builds, at render distance 6, min FPS is around 120, and at render distance 10, min FPS is around 25. So my guess is that there will be similar % improvements with release builds, which would result in really high FPS (maybe ~400 min FPS at render distance 6 and ~50-60 at render distance 10).